Support native resolution of the SSH config
Trying to resolve the config for a given host manually may run into
incompatibilities between the parsing logic of the extension and actual
SSH. For example, the extension does not currently apply hostname
patterns the same way as OpenSSH does. This may result in failing to
connect to the host, if, e.g., the missed pattern was to add a ProxyJump
field.

While bugs like this can be fixed, it is convenient to just defer to SSH
itself when resolving the config. This makes it much more robust,
especially against future changes by SSH.

The native resolver is marked experimental and opt in. Possibly other
clients besides OpenSSH have different arguments to print the resolved
config (or don't at all).

Signed-off-by: Benjamin Gray <bgray@linux.ibm.com>